fix: fix clickhouse stats

+++ internal/service/analitics/models.go
@@ -6,7 +6,7 @@ type Message struct {
 	CreatedAt time.Time `json:"created_at"`
 
 	TgChatID int64  `json:"tg_chat_id"`
-	TgId     int    `json:"tg_id"`
+	TgId     int64  `json:"tg_id"`
 	TgUserId int64  `json:"tg_user_id"`
 	Text     string `json:"text"`
 }
